requirements.txt
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
eval_type_backport==0.2.0
imageio==2.36.0
lazy_loader==0.4
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydicom==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.1
simsimd==5.9.11
six==1.16.0
stringzilla==3.10.7
tdqm==0.0.1
tifffile==2024.9.20
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
xlrd==2.0.1
